chore(frameworks): add mobile view
